+++ cmd/sync.go
@@ -51,6 +51,7 @@ type rawSyncCmdArgs struct {
 	legacyInclude         string // for warning messages only
 	legacyExclude         string // for warning messages only
 
+	preserveNTFSACLs    bool
 	followSymlinks      bool
 	putMd5              bool
 	md5ValidationOption string
@@ -176,6 +177,11 @@ func (raw *rawSyncCmdArgs) cook() (cookedSyncCmdArgs, error) {
 		return cooked, err
 	}
 
+	cooked.preserveNTFSACLs = raw.preserveNTFSACLs
+	if err = validatePreserveNTFSACLs(cooked.preserveNTFSACLs, cooked.fromTo); err != nil {
+		return cooked, err
+	}
+
 	cooked.putMd5 = raw.putMd5
 	if err = validatePutMd5(cooked.putMd5, cooked.fromTo); err != nil {
 		return cooked, err
@@ -231,6 +237,7 @@ type cookedSyncCmdArgs struct {
 	excludeFileAttributes []string
 
 	// options
+	preserveNTFSACLs    bool
 	putMd5              bool
 	md5ValidationOption common.HashValidationOption
 	blockSize           uint32
@@ -562,6 +569,7 @@ func init() {
 
 	rootCmd.AddCommand(syncCmd)
 	syncCmd.PersistentFlags().BoolVar(&raw.recursive, "recursive", true, "True by default, look into sub-directories recursively when syncing between directories. (default true).")
+	syncCmd.PersistentFlags().BoolVar(&raw.preserveNTFSACLs, "preserve-ntfs-acls", false, "False by default. Preserves NTFS ACLs between aware resources (Windows and Azure Files)")
 	syncCmd.PersistentFlags().Float64Var(&raw.blockSizeMB, "block-size-mb", 0, "Use this block size (specified in MiB) when uploading to Azure Storage or downloading from Azure Storage. Default is automatically calculated based on file size. Decimal fractions are allowed (For example: 0.25).")
 	syncCmd.PersistentFlags().StringVar(&raw.include, "include-pattern", "", "Include only files where the name matches the pattern list. For example: *.jpg;*.pdf;exactName")
 	syncCmd.PersistentFlags().StringVar(&raw.exclude, "exclude-pattern", "", "Exclude files where the name matches the pattern list. For example: *.jpg;*.pdf;exactName")