+++ ste/JobPartPlan.go
@@ -14,7 +14,7 @@ import (
 // dataSchemaVersion defines the data schema version of JobPart order files supported by
 // current version of azcopy
 // To be Incremented every time when we release azcopy with changed dataSchema
-const DataSchemaVersion common.Version = 10
+const DataSchemaVersion common.Version = 11
 
 const (
 	CustomHeaderMaxBytes = 256
@@ -58,6 +58,7 @@ type JobPartPlanHeader struct {
 	DstBlobData           JobPartPlanDstBlob  // Additional data for blob destinations
 	DstLocalData          JobPartPlanDstLocal // Additional data for local destinations
 
+	PreserveNTFSACLs bool
 	// S2SGetPropertiesInBackend represents whether to enable get S3 objects' or Azure files' properties during s2s copy in backend.
 	S2SGetPropertiesInBackend bool
 	// S2SSourceChangeValidation represents whether user wants to check if source has changed after enumerating.